feat(infra): create separate service for json serialization and deserialization
